requirements.txt
asgiref==3.6.0
captcha==0.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.5
django-allauth==0.52.0
django-crispy-forms==1.14.0
django-jet-reboot==1.3.3
django-paypal==2.0
django-recaptcha==3.0.0
django-tinymce==3.5.0
fontawesomefree==6.2.1
google-api-python-client==1.4.1
httplib2==0.21.0
idna==3.4
oauth2client==4.1.3
oauthlib==3.2.2
Pillow==9.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
stripe==5.1.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
